requirements.txt
requests
python-dotenv
psycopg2-binary
minio
pyspark
scikit-learn
implicit
flask
boto3
pandas
pyarrow
fastparquet